requirements.txt
# Core dependencies
aiofiles==24.1.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
requests==2.32.3
urllib3==2.4.0

# Azure Document Intelligence for PDF processing
azure-ai-documentintelligence>=1.0.2

# Database
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
pgvector==0.2.4

# LangChain for LLM operations
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-openai==0.3.18
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langgraph==0.2.55
langgraph-cli[inmem]
langsmith==0.3.42

# OpenAI
openai==1.82.0
tiktoken==0.9.0

# Streamlit UI
streamlit==1.45.1
altair==5.5.0
blinker==1.9.0
cachetools==5.5.2
click==8.2.1
gitdb==4.0.12
GitPython==3.1.44
importlib-metadata==8.5.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==1.41.0
packaging==23.2
pillow==11.2.1
protobuf==5.26.1
pyarrow==20.0.0
pydeck==0.9.1
rich==13.7.1
smmap==5.0.2
toml==0.10.2
tornado==6.5.1
tzlocal==5.3.1
watchdog==6.0.0

# Japanese text processing
sudachipy==0.6.8
sudachidict-core>=20240716
rank_bm25>=0.2.2

# Clustering for domain classification
hdbscan>=0.8.29
umap-learn>=0.5.3

# Data processing
numpy==1.26.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
openpyxl==3.1.5

# Configuration
python-dotenv==1.1.0
PyYAML==6.0.2

# Pydantic for data validation
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2

# JSON processing
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
orjson==3.10.18
dataclasses-json==0.6.7

# Async support
aiohttp==3.12.2
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
frozenlist==1.6.0
multidict==6.4.4
propcache==0.3.1
yarl==1.20.0

# HTTP clients
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0

# Utilities
attrs==25.3.0
backoff==2.2.1
colorama==0.4.6
distro==1.9.0
greenlet==3.2.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
marshmallow==3.26.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
referencing==0.36.2
regex==2024.11.6
requests-toolbelt==1.0.0
rpds-py==0.25.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==8.5.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.2
wrapt==1.17.2

# Optional: Graph visualization (can be removed if not needed)
pyvis==0.3.2
plotly==6.1.2

# Hugging Face for local LLM support
transformers==4.47.1
torch==2.5.1
accelerate==1.2.1
sentence-transformers==3.3.1
bitsandbytes==0.45.0